integrate pr #162 changes with existing global config
