--- src/main/resources/io/jenkins/plugins/bitbucketpushandpullrequest/configuration/BitBucketPPRNotifyConfiguration/config.jelly
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
-         xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <f:section title="${%BitBucket Push and Pull Request}">
-        <f:entry title="${%Send build status to BitBucket}" field="notifyBitBucket">
-            <f:checkbox default="true" />
-        </f:entry>
-    </f:section>
-</j:jelly>
